Add osc.sh (clipboard tool)
This comes from

https://chromium.googlesource.com/apps/libapps/+/HEAD/nassh/docs/FAQ.md#Is-OSC-52-aka-clipboard-operations_supported

OSC-52 is a protocol for sending results to the clipboard. It is useful
when connecting to a system via ssh, and then wanting something on the
system clipboard.

Usage is to copy stdin (or a string on the command line) to the
clipboard.

This works with the Chrome SSH extension, which I use extensively to
connect to linux systems.