Add aliases for dtach
I now often use dtach as a minimalist replacement for tmux. This creates
two aliases for dtach to connect to a lhs and rhs session. This is part
oa workflow where I have one thing on the left window, and another on
the right.

Reconnecting to a failed session and typing lhs is sufficient to restore
that session.